backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.10.4
pydantic-settings==2.7.1
python-multipart==0.0.20
supabase==2.10.0
pillow==10.4.0
pillow-heif==0.21.0
imagehash==4.3.1
openai==1.59.7
langchain==0.3.0
langchain-openai==0.2.0
langchain-core==0.3.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.27.0
